backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
google-auth==2.36.0
openai==1.51.0
httpx==0.27.2
aiohttp==3.10.10
requests==2.32.3
SpeechRecognition==3.10.0
pyttsx3==2.90
bleach==6.1.0
python-slugify==8.0.4
pyyaml==6.0.2
slowapi==0.1.9
starlette-context==0.3.6
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2
